Updated Makefile - Seperated Expansion file - Added make valrun
