+++ minishell/Makefile
@@ -6,7 +6,7 @@
 #    By: zali <zali@student.42.fr>                  +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2025/07/11 09:44:49 by vloureir          #+#    #+#              #
-#    Updated: 2025/07/25 13:57:45 by zali             ###   ########.fr        #
+#    Updated: 2025/07/27 17:29:46 by zali             ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -16,9 +16,9 @@
 
 NAME := minishell
 RUN_CMD := $(wildcard runcmd/*.c)
+EXPAN_CMD := $(wildcard expansion/*.c)
 SRC := $(wildcard *.c)
-
-OBJ := $(SRC:.c=.o) ${RUN_CMD:.c=.o}
+OBJ := $(SRC:.c=.o) $(RUN_CMD:.c=.o) $(EXPAN_CMD:.c=.o)
 
 VAL := --leak-check=full --show-leak-kinds=all --track-origins=yes --show-reachable=yes --suppressions=readline.supp
 
@@ -27,6 +27,9 @@ CFLAGS := -Wall -Werror -Wextra -g
 LFLAG := -lreadline -Llibft -lft
 
 all: $(NAME)
+
+valrun: $(NAME)
+	valgrind $(VAL) ./$(NAME)
 
 $(NAME): $(OBJ)
 	$(MAKE) -C ./libft
