+++ minishell/runcmd/parse.c
@@ -6,7 +6,7 @@
 /*   By: zali <zali@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/20 16:36:13 by zali              #+#    #+#             */
-/*   Updated: 2025/07/26 19:39:17 by zali             ###   ########.fr       */
+/*   Updated: 2025/07/27 14:48:00 by zali             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
